+++ GatheringPathRenderer/Configuration.cs
@@ -5,4 +5,5 @@ internal sealed class Configuration : IPluginConfiguration
 {
     public string AuthorName { get; set; } = "?";
     public int Version { get; set; } = 1;
+    public bool ShowOverlay { get; set; } = true;
 }