+++ Questionable/Windows/JournalComponents/QuestJournalComponent.cs
@@ -132,7 +132,7 @@ internal sealed class QuestJournalComponent
         ImGui.TableNextColumn();
 
         string genreName = filter.Genre.Name;
-        if (questRegistry.TryGetQuest(filter.Quests[0].QuestId, out Quest? q))
+        if (filter.Genre.Id != JournalData.InstantQuests && questRegistry.TryGetQuest(filter.Quests[0].QuestId, out Quest? q))
         {
             RedoIndex redoIndex = redoUtil.GetChapter(q.Id.Value);
             if (redoIndex.Index != -1)