+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-- Bug fix: title bar minimize button -alydev
+- Feature: added "instant" quests (So You Want To Be A Questionable User) to Journal Progress > Other Quests > Special Quests > Instant Quests -alydev
+  - the "Instant Quests" label can be right clicked to select "Add all to priority quests", this will let users run all available one-step quasi-quests in one go (in theory!)
+  - bug fix: don't apply NG+ label to Instant Quests category; sort by quest ID -alydev
+- Feature: replaced button labels (visible on hover) to quick access buttons -alydev